fix :start_date nullification on update
When a user was updating a mission's attributes other than :start_date
and :end_date, these date attributes were nullified as an empty field
was sent in the update request.
Now, these fields fetch their own value if it exists, and nullification
after update is prevented on :start_date to prevent the event from disappearing from
the calendar.