fix(migrations): needed to bump thredded, then sassc, then puma
`rails db:migrate` would crash on thredded forum migration. The
following changes were needed:
- bump thredded, which needed sassc, which needed sprockets 4
- bump puma to v6, for Rack 3 compatibility, which was needed for the
system tests to launch with Capybara
- update the Member model to declare the enum type through `attribute`,
because when the thredded migrations start, the Member model is then
loaded. Since Rails 7.1, enums must be backed by the `attribute` method
or a database column, and at the time when the migration accesses the
Member model, that column `cash_register_proficiency` does not exist yet
(this raised error is reverted on Rails 7.2)